move biosimulator tests to separate file
